A lot of things changed, should've commited before...Anyway...Updates!
SO, I've modified Fog Light s to look better, but also included a new function : change_radius. This function takes for options : the new radius, the modification speed (in-pixels), and the over_radius (this option defines a radius of more exaggerated  size, this allows for a small "bobbing" for the fog_light and to make it more "lively", but this feature is still buggy and doesn't fully work).

I've also added the possibility to make the engine "see" meta-tiles and others as game objects, thus allowing to make primitive background and rooms.
No proper map systems YET.

I've also added layers, lemme explain : when you call the Engine class function add_object, it'll require one more option, a layer option. It can range from 0 to infinite, negative values are technically supported, but actually relates to the positive value equal to the number of layers plus your negative value, so true_value = layer_num  + neg_value. So this can end up in weird situations if you use negative values.

I've also added the possibility to change the fog's color. The variable is located in the Engine class under the name of fog_color, you need to change it's value directly though, there is no function for that.

I think that's about all the changes in the commit for now. At the moment we still need:
 -Collision resolution
 -Map loading system
 -Map editor
 -Sound integration (no custom data storing though, we're keeping .ogg)